resolving comments on PR

+++ src/main/java/com/gladguys/alucoapi/controllers/ClassController.java
@@ -1,6 +1,5 @@
 package com.gladguys.alucoapi.controllers;
 
-import com.gladguys.alucoapi.entities.Class;
 import com.gladguys.alucoapi.entities.StudentWrapper;
 import com.gladguys.alucoapi.entities.dto.ClassDTO;
 import com.gladguys.alucoapi.security.jwt.JwtTokenUtil;
@@ -17,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
+
 import java.util.Set;
 
 @RestController
@@ -34,6 +34,8 @@ public class ClassController {
 	@GetMapping("/teacher/{teacherId}")
 	public ResponseEntity<Set<ClassDTO>> getAllByTeacher(@PathVariable("teacherId") Long teacherId) {
 		try {
+			if(teacherId == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+
 			Set<ClassDTO> classes = this.classService.getAllByTeacher(teacherId);
 			return ResponseEntity.ok(classes);
 
@@ -55,16 +57,25 @@ public class ClassController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<ClassDTO> getById(@PathVariable("id") Long id) throws Exception {
-		ClassDTO classFound = this.classService.getById(id);
-		return ResponseEntity.ok(classFound);
+	public ResponseEntity<ClassDTO> getById(@PathVariable("id") Long id) {
+		try {
+			if(id == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+
+			ClassDTO classFound = this.classService.getById(id);
+			return ResponseEntity.ok(classFound);
+
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+		}
 	}
 
 	@PostMapping
 	public ResponseEntity<ClassDTO> save(@RequestBody ClassDTO dto) {
 		try {
+			if(dto == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 			ClassDTO classDTO = this.classService.saveOrUpdate(dto);
-			return ResponseEntity.ok(classDTO);
+
+			return ResponseEntity.status(HttpStatus.CREATED).body(classDTO);
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
 		}
@@ -74,13 +85,17 @@ public class ClassController {
 	public ResponseEntity<ClassDTO> update(@RequestBody ClassDTO dto, HttpServletRequest request) {
 
 		try {
+			if(dto == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+			else if (dto.getId() == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+
 			boolean exists = this.classService.exists(dto.getId());
 			if (!exists) return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
 
 			Long teacherId = jwtTokenUtil.getTeacherIdFromToken(request).longValue();
 			dto.setTeacherId(teacherId);
 
-			return ResponseEntity.ok(this.classService.saveOrUpdate(dto));
+			ClassDTO classSaved = this.classService.saveOrUpdate(dto);
+			return ResponseEntity.ok(classSaved);
 
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
@@ -90,8 +105,12 @@ public class ClassController {
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> delete(@PathVariable("id") Long id) {
 		try {
+			boolean exists = this.classService.exists(id);
+			if (!exists) return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+
 			this.classService.deleteById(id);
-			return ResponseEntity.status(HttpStatus.OK).body("turma removida com sucesso.");
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
+
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
 		}
@@ -101,11 +120,11 @@ public class ClassController {
 	@PostMapping("/{id}/students")
 	public ResponseEntity saveStudentsForClass(@PathVariable("id") Long id, @RequestBody StudentWrapper wrapper) {
 		try {
-			if(wrapper == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
-					"não foi encontrado a lista de estudantes a adicionar");
+			if(id == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+			if(wrapper == null || wrapper.getStudentDTOS().isEmpty()) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 
 			this.classService.addStudentsIntoClass(wrapper.getStudentDTOS(), id);
-			return ResponseEntity.ok("estudantes adicionados a classe ");
+			return ResponseEntity.status(HttpStatus.CREATED).body(null);
 
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
